add initial structure for wtf nodes
wtf users must create a wtfconfig.py file to define the nodes in the test
setup.  This file is imported by the tests, which command the nodes to do
certain things.  The first wtfconfig.py file found in the python import path
will be used.  See platform/mwl8k/wtfconfig.py for a sample config.