tune up top-level test runner
Specifically:
-- It's now called wtf.py.  Neat.

-- README is updated.  Have a look.

-- Make config file dumber and config class smarter by having it
   determine the node lists from a single input list.

-- You can now specify a config file on the command line.

-- Added back the delay to wait for slow APs to start up in the
   ap_sta test.  This allows the microclient jr AP to work again.

I eliminated support for specifying multiple suites in the config
files.  This could be added back if we want, but I took it out
for now.

I also eliminated our ability to specify multiple configs in the
config file.  This could also be added back, but we weren't using
it, and it was kinda messy.