add simple p2p find test
p2p (wifi direct) is something that we are testing on some
projects.  So introduce a test suite and a base class.  Also add
implementations for wpa_supplicant's p2p functionality.  Note
that functionality in wpa_supplicant is under heavy development.
So these tests may have to be dusted off for a while before they
stabilize.