add force_driver_reload option to mvdroid p2p
the force_driver_reload option to the Mvdroid constructor forces
a full reload of the underlying driver at node start/stop time
instead of node init/shutdown time.  Currently, this option must
be set to true to prevent the state of one test influencing
another test.  Clearly, in a real system you would not want to
reload the drivers for each wifi-direct use case.  This is a bug
in mvdroid.  Once it's fixed, we can stop using
force_driver_reload=True.