initial tshark test
With this patch we establish a means to test 80211 packet
construction and dumping using tshark.  Specifically, we use
scapy to create a packet.  Then we pass it through tshark and
dump the output as xml.  Then we parse the xml with lxml and
ensure that the input packet is the same as the output packet.

With contributions from Johan Nestaas