arduino mc200 build, flash, and test output
two changes that were not in this test are as follows:
* allow a comm object to be stored inside of wtf.config
normally the comm is stored inside of the linuxNode but this test
only has access to the Serial object so the comm was needed from the
wtf.config
* add data field to wtf.conf object
allows for a dictionary to be passed from the config. The data must
be validated inside of test. This lets computer specific data (like
the path to an executable) to be passed from the config file.