hostapd: update to conform to new platformOps way of life
More data is now stored inside of the iface, which is now a list of
objects, so anywhere self.iface was called before, is now
self.iface[0].name since before the iface was just "wlan0". Change the
arguments passed to Hostapd.__init__ to remove driver (now inside of
iface) and add ops. Lastly put a line break before writing text inside
of the textwrap.dedent to avoid whitespace mangling the hostapd.conf